// payload-types.ts
export interface Experience {
  id: number;
  company: string;
  position: string;
  startDate: string;
  /**
   * Kosongkan jika masih berlangsung
   */
  endDate?: string | null;
  description: string;
  badges?: string[] | null;
  updatedAt: string;
  createdAt: string;
}